% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CallStop}
\alias{CallStop}
\title{Stop on error and print error message.}
\usage{
CallStop(errstr)
}
\arguments{
\item{errstr}{is a string which hold the error message to be printed before
stopping execution}
}
\description{
This function stops and prints an error message which has been
  passed to it. It does not have a return value.
}

