% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CheckHerzParams}
\alias{CheckHerzParams}
\title{Check HErZ parameter names.}
\usage{
CheckHerzParams(herz.param, herz.profile)
}
\arguments{
\item{herz.param}{a string of length n which holds n parameter names.}

\item{herz.profile}{boolean which determines whether to check for the complete
HErZ profile or only the two levels 10m and 116m.}
}
\description{
HErZ parameter names are checked which is especially important if
  the complete profile is passed. Within the package the names are expected to
  follow a certain structure: "windspeed_xxxm", where xxx equals the height of
  the model level and is expected to be decreasing, i.e., 258m, 178m, 116m, 69m,
  35m, 10m. This function does not have a return value.
}

