% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CheckLindParams}
\alias{CheckLindParams}
\title{Check parameter names of Lindenberg tower measurements.}
\usage{
CheckLindParams(lind.param)
}
\arguments{
\item{lind.param}{a string of length n which holds n parameter names.}
}
\description{
Lindenberg parameter names are checked. Within the package the names
  are expected to follow a certain structure: "windspeed_xxxm", whereas xxx
  equals the height of the measurement level and is expected to be increasing,
  i.e., 10m, 20m, 40m, 80m, 98m. This function does not have a return value.
}

