% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ExtractERAxts}
\alias{ExtractERAxts}
\title{Extract an extended time series off a 2D ERA data (ERA20C, ERA-I)
corresponding to a provided point (lon, lat) location.}
\usage{
ExtractERAxts(data, time.vals, lon, lat, tsstart, tsend, stat.lon, stat.lat)
}
\arguments{
\item{data}{(reanalysis=) data set from which to extract the time series at the
provided station location}

\item{time.vals}{time period covered by the data set}

\item{lon}{longitude grid points of the data set}

\item{lat}{latitude grid points of the data set}

\item{tsstart}{character string of the format (YYYY,M) for the newly to be set
start date of the data set}

\item{tsend}{character string as above for the end date}

\item{stat.lon}{station longitude}

\item{stat.lat}{station latitude}
}
\value{
era.xts is the time series extracted off the data set at the
  pixel corresponding to the provided station location.
}
\description{
\code{ExtractERAxts} extracts the nearest pixel of a provided data
  set (reanalysis) given the station location (lon, lat).
}

