% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ExtractHErZxts}
\alias{ExtractHErZxts}
\title{Extract an extended time series off 2D HErZ data corresponding to a
provided point (lon, lat) location.}
\usage{
ExtractHErZxts(herz.data, time.vals, herz.lon, herz.lat, tsstart, tsend,
  stat.lon, stat.lat)
}
\arguments{
\item{herz.data}{is the HErZ data set from which to extract the time series at
the station location}

\item{time.vals}{time period covered by the HErZ data set}

\item{herz.lon}{longitude grid points of the HErZ data set}

\item{herz.lat}{latitude grid points of the HErZ data set}

\item{tsstart}{character string of the format (YYYY,M) of the start date of the
HErZ data set time series}

\item{tsend}{character string as above of the end date of the HErZ data set}

\item{stat.lon}{station longitude}

\item{stat.lat}{station latitude}
}
\value{
herz.xts[timestr]is the time series extracted off the HErZ reanalysis at
  the pixel corresponding to the provided station location
  Extract nearest HErZ pixel corresponding to station lon, lat calling Uni Bonn
  function getNearest.
}
\description{
\code{ExtractHErZxts} extracts the time series of the HErZ pixel
  corresponding to the station location. HErZ data come in an irregular polar
  projection which calls for a function to extract the correct pixel off that grid.
}

