% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ExtractStationData}
\alias{ExtractStationData}
\title{Extract and align wind station data to reanalysis data.
  \code{ExtractStationData} reads wind speed values and dates from station data
  and creates a gap filled (NA), hence continuous, extended time series. Monthly
  or daily mean time series are calculated which span the time period of the
  longest stretching reanalysis data set. The time steps of the monthly time
  series is given with precision of months, those of daily time series with
  precision of days. Parts taken from get.timeseries.R of GetPlotsFromFtp and off
  website http://bocoup.com/weblog/padding-time-series-with-r/}
\usage{
ExtractStationData(station.data, era20c.tsstart, era20c.tsend, eraI.tsstart,
  eraI.tsend, herz.tsstart, herz.tsend, ana.time.res, station.daily)
}
\arguments{
\item{station.data}{is an extended time series holding monthly mean station wind
data}

\item{era20c.tsstart}{is a charcter string of the start date of the ERA20C data
of the format c(YYYY,MM)}

\item{era20c.tsend}{is the same as above for the end date}

\item{eraI.tsstart}{is the start date of ERA-Interim data}

\item{eraI.tsend}{is the end date of ERA-Interim data}

\item{herz.tsstart}{is the start date of HErZ data}

\item{herz.tsend}{is the end date of the HErZ data}

\item{ana.time.res}{named list holding parameters monthly="monthly",
daily="daily", hourly="hourly", and time.res= to determine the time resolution
of the data to be read.}

\item{station.daily}{boolean parameter to specify whether station data is
aggregated in hourly (F) or daily (T) time steps.}
}
\value{
MM.station[timestr] is the potentially gab filled monthly mean station
  data spanning the time period of the longest ranging reanalysis time series.
}
\description{
Extract and align wind station data to reanalysis data.
  \code{ExtractStationData} reads wind speed values and dates from station data
  and creates a gap filled (NA), hence continuous, extended time series. Monthly
  or daily mean time series are calculated which span the time period of the
  longest stretching reanalysis data set. The time steps of the monthly time
  series is given with precision of months, those of daily time series with
  precision of days. Parts taken from get.timeseries.R of GetPlotsFromFtp and off
  website http://bocoup.com/weblog/padding-time-series-with-r/
}
\note{
needs to add functionality to calculate daily or monthly values only if
  there are enough hourly values available; it needsd to be decided what enough
  is.
}

