% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ExtractTowerData}
\alias{ExtractTowerData}
\title{Extract tower measurement data.}
\usage{
ExtractTowerData(file.name, para.name, era.mon = FALSE)
}
\arguments{
\item{file.name}{character string holding the file name of the file to be read}

\item{para.name}{character string holding the parameter name(s) to be read}

\item{era.mon}{optional boolean which decides whether to format the time axis
as daily or monthly.}
}
\value{
The return value is a named list with the names of the parameter names
  holding extended time series of that parameter.
}
\description{
This function reads one or more variable name(s) off a netCDF file
  and formats the time axis. A list named by the variable names is returned
  which holds extended time series for the complete time series of the variable
  name specified (or all).
}

