% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.data.download}
\alias{get.data.download}
\title{Download data from the DWD-ftp server.}
\usage{
get.data.download(metadata = NULL)
}
\arguments{
\item{metadata}{list. Argument retrieved
from \code{\link{get.metadata}}. If empty, the
user will be asked to enter the metadata
manually through the R-console.}
}
\description{
\code{get.data.download}
  Data is downloaded from the DWD-ftp server
  according to the metadata information saved in
  the output argument \code{metadata} retrieved from
  \code{\link{get.metadata}}. No object is returned.
}

