% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GetCorrXts}
\alias{GetCorrXts}
\title{Correlation between two (extended) time series.}
\usage{
GetCorrXts(era20c, eraI, herz, stat)
}
\arguments{
\item{era20c}{extended time series of an ERA20C pixel corresponding locally to
the station time series}

\item{eraI}{same as above for ERA-Interim}

\item{herz}{same as above for HErZ}

\item{stat}{same as above for the station values}
}
\value{
Return a named list (c.20c.I=,c.20c.H=,c.20c.S=,c.I.H=,c.I.S=,c.H.S=)
  holding all possible perturbations of correlations between the provided four
  extended time series.
}
\description{
\code{GetCorrXts} calculates the correlation between two extended
  time series if it is possible, i.e., if they span the same time period.
  Otherwise return NA. This function is hardcoded to handle all combinations
  between four provided extended time series.
}

