% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GetData}
\alias{GetData}
\title{Download data from the DWD-ftp server.}
\usage{
GetData(metadata = NULL, verbose.DWD = TRUE)
}
\arguments{
\item{metadata}{list; argument retrieved from \code{\link{GetMetadata}}. If
empty, the user will be asked to enter the metadata manually through the
R-console.}

\item{verbose.DWD}{boolean; determines whether to tell what's going on.}
}
\value{
\code{data1} data.frame. Contains the data that correspond to the given
  metadata.
}
\description{
\code{GetData}
  Allows downloading data from the DWD-ftp server according to the metadata
  information saved in the output argument \code{metadata} retrieved from
  \code{\link{GetMetadata}}. The data is then unzipped and extracted for hourly
  or daily values as specified.
}
\examples{
# If no input argument is known:
  \dontrun{
  data1 <- GetData() # R will ask the user through the
                      # console to enter each argument
  # If the input argument is known
  data1 <- GetData(metadata)
  }
}

