% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GetDataDownload}
\alias{GetDataDownload}
\title{Download data from the DWD-ftp server.}
\usage{
GetDataDownload(metadata = NULL, verbose.DWD = TRUE)
}
\arguments{
\item{metadata}{list. Argument retrieved
from \code{\link{GetMetadata}}. If empty, the
user will be asked to enter the metadata
manually through the R-console.}

\item{verbose.DWD}{boolean; determines whether to tell what's going on.}
}
\description{
\code{GetDataDownload}
  Data is downloaded from the DWD-ftp server
  according to the metadata information saved in
  the output argument \code{metadata} retrieved from
  \code{\link{GetMetadata}}. No object is returned.
}

