% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GetLonLatIdx}
\alias{GetLonLatIdx}
\title{Extract longitude and latitude index off a lon, lat regular grid.}
\usage{
GetLonLatIdx(fname, point.lon, point.lat, grid.lon = NULL, grid.lat = NULL)
}
\arguments{
\item{fname}{string of the file name to read}

\item{point.lon}{longitude point which to match and extract off the grid}

\item{point.lat}{latitude point as above}

\item{grid.lon}{optional parameter of the grid longitude values. If the grid#
ongitude is a vector and can be read from nc file this is done and the grid.lon
parameter needs to be set to NULL. The default is to set grid.lon to NULL.}

\item{grid.lat}{same as above for the grid latitude values.}
}
\value{
Return a named lilst of (lonidx=,latidx=) longitude and latitude index.
}
\description{
\code{GetLonLatIdx} extracts the longitude and latitude index of
a point off a netCDF file which holds data on a regular grid with longitude and
latitude values (as vectors) stored.
}

