% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GetNearestIdx}
\alias{GetNearestIdx}
\title{Get nearest neighbor index of a point within a vector.}
\usage{
GetNearestIdx(vec, num)
}
\arguments{
\item{vec}{vector of regularly spaced values}

\item{num}{numeric to find the nearest index within the provided vector.}
}
\value{
idx the nearest index of \code{num} within \code{vec}.
}
\description{
\code{get.lonidx} calculates the nearest neighbor of a point within
a vector. Here, it is used to find the index within a regular lon, lat grid of a
specified lon, lat point. This function needs to be executed separately for each
index. The value of the point needs to be within the range of the vector.
}

