% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GetSeasonalXts}
\alias{GetSeasonalXts}
\title{Extract seasonal time series off a monthly time series.}
\usage{
GetSeasonalXts(vals.xts)
}
\arguments{
\item{vals.xts}{is the input extended time series which is divided into seasonal
(quarterly) xts.}
}
\value{
A named list returning a winter, spring, summer, and autumn extended
 time series, (winter.xts=,spring.xts=,summer.xts=,autumn.xts=),  which was split
 off the input xts.
}
\description{
This fucntion extracts a seasonal time series off a monthly time
  series by using the function split.xts. However, it only splits into calendar
  quarters, not into meteorological seasons. The return values are the separate
  seasonal (quarterly) time series.
}
\note{
could be enhanced by manually splitting into meteorol. seasons DJF, MAM,
  JJA, SON instead of using the generic split 'f="quarters"'.
}

