% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GetYlims}
\alias{GetYlims}
\title{Determine y-axis limits from data.}
\usage{
GetYlims(xts1, xts2, xts3, xts4)
}
\arguments{
\item{xts1,xts2,xts3,xts4}{extended time series from which to determine the low
and high range of the y-axis limits.}
}
\value{
Return a named list (yll=,ylh=) of the lower and high bound of the y-axis
  limits yliml and ylimh.
}
\description{
\code{GetYlims} determinies the low and high y-axis limits from four
  different time series.
  This needs to be enhanced so that not all of those time series need to be
  available. Something like present= in FORTRAN.
}

