% Generated by roxygen2 (4.0.2): do not edit by hand
\name{NormVals}
\alias{NormVals}
\title{Calculate the norm of a value given the the upper and lower bound.}
\usage{
NormVals(value, min.val, max.val)
}
\arguments{
\item{value}{is the value of which the norm is to be calcualted. Needs to satisfay:
min.value =< value =< max.val.}

\item{min.val,max.val}{are both the lower and upper bound}
}
\value{
is the norm value.
}
\description{
Given the upper and lower bound calculate the norm value.
}

