% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Plot100mEraHerz}
\alias{Plot100mEraHerz}
\title{Compare 100m wind speed of ERA20C and HErZ pixel by pixel.}
\usage{
Plot100mEraHerz(Era20cXts, HerzXts, titname, outdir, fname, width, height)
}
\arguments{
\item{Era20cXts}{monthly mean extended time series of a ERA20C pixel}

\item{HerzXts}{same as above for HErZ}

\item{titname}{string of the plot title name}

\item{outdir}{string of the output directory into which the plot is saved}

\item{fname}{string of the file name of the plot}

\item{width,height}{of the plot in inches}
}
\description{
\code{Plot100mEraHerz} compares the 100m wind speed of the ERA20C
  global reanalysis with the 116m wind speed of the HErZ regional reanalysis. As of
  now THIS FUNCTION IS NOT YET FULLY FUNCTIONAL !!!
  This function will perform a pixel wise comparison to fit into the logic of this
  package. Later an areal comparison is envisaged.
  Scatter plots, PDFs, histograms, PDFscore, etc should all be considered as
  verification scores.
}

