% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PlotHistogramsTower}
\alias{PlotHistogramsTower}
\title{Plot histrograms for tower measurements.}
\usage{
PlotHistogramsTower(outdir, fname, era.monthly, tower.obj)
}
\arguments{
\item{outdir}{is a string containing the output path of the plot}

\item{fname}{is a string of the file name of the plot file}

\item{era.monthly}{is an optional boolean which determines whether data passed
is monthly (T) or daily (F) data. The default value is to use daily data
(era.monthly=FALSE).}

\item{tower.df}{is a data frame holding the data of tower measurements and
corresponding reanalysis data}
}
\description{
Plot histograms for the Lindenberg, Cabauw, Fino1 and Fino2 tower
  tower measurements. Histograms are plotted for 10m and 100m heights; seperately
  and in comparison with reanalysis data.
}

