% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotLegendStats}
\alias{plotLegendStats}
\title{Plot legend with statistics for histogram plot.}
\usage{
plotLegendStats(xlims, vals)
}
\arguments{
\item{xlims}{the limits of the x-axis in the form c(low,high)}

\item{vals}{vector of the plotted values of which the statistics are calculated}
}
\description{
Plot the legend for histogram plots with statistics (mean, median,
  etc) and weibull parameters which are calculated.
}

