% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PlotMultiPanel}
\alias{PlotMultiPanel}
\title{Plot a 2-by-2 multi panel plot.}
\usage{
PlotMultiPanel(outdir, fname, titname, Era20c, EraI, Herz, Stat, length.plot,
  era.months, plot.diff = FALSE)
}
\arguments{
\item{outdir}{string of the output directory into which the plot is saved}

\item{fname}{string of the file name of the plot}

\item{titname}{string of the plot title name}

\item{Era20c}{extended time series of monthly or seasonal ERA20C data}

\item{EraI}{extended time series of momthly or seasonal ERA-Interim data}

\item{Herz}{extended time series of monthly or seasonal HErZ data}

\item{Stat}{extended time series of monthly or seasonal station data}

\item{length.plot}{a list holding integers for the months or seasons to plot.
It holds 1, .., 12 specifying the months, or 1,..,4 specifying the seasons.}

\item{era.months}{boolean which specifies whether monthly or seasonal data
shall be plotted.}

\item{plot.diff}{optional boolean which determines whether to plot the difference
of the reanalyses to the station data or the absolute values of each time
series. When plotting the difference, only three panels will be plotted. The
default value is to plot the absolute values of each time series
(plot.diff=FALSE).}
}
\description{
This is supposed to be a rather generic routine to plot a 2-by-2
  multi panel plot. So far, it is used to plot the four different data sources
  (ERA20C, ERA-Interim, HErZ, station data) for monthly and seasonal values. The
  plot is saved into a pdf file, and the function does not have a return value.
}

