% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PlotPDFScore}
\alias{PlotPDFScore}
\title{Calculate the S_score as described in Mayer et al., 2015.}
\usage{
PlotPDFScore(era.xts, station.xts, outdir, fname, titname, ana.time.res)
}
\arguments{
\item{era.xts}{is the reanalysis time series of one pixel}

\item{station.xts}{is the corresponding station time series}

\item{outdir}{is a string containing the output path of the plot}

\item{fname}{is a string holding file name of the pdf plot to be created.}

\item{titname}{is a string containig the title name of the plot}

\item{ana.time.res}{named list holding parameters monthly="monthly",
daily="daily", hourly="hourly", and time.res= to determine the time resolution
of the data to be read.}
}
\description{
Mayer et al., 2015: "Identifying added value in high-resolution
  climate simulations over Scandinavia" describe the S_score based on the
  examination of two PDFs and checking their overlap. Here, these two PDFs are
  built off one pixel of a reanalysis time series and the corresponding station
  time series. Here, the S_score is plotted as January through December values,
  displaying the annual cycle.  The plot is saved in pdf format and there is no
  return value.
}

