% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PlotStationEra}
\alias{PlotStationEra}
\title{Plot station measurements together with ERA20C, ERA-I, and HErZ.}
\usage{
PlotStationEra(s.obj, titname, outdir, fname, anomaly = FALSE)
}
\arguments{
\item{s.obj}{is a ClimObject which holds the station data and corresponding
reanalysis data at the location of the station data.}

\item{titname}{string of the plot title name}

\item{outdir}{string of the output directory into which the plot is saved}

\item{fname}{is a string holding file name of the pdf plot to be created.}

\item{anomaly}{is an optional parameter which determines whether to plot anomalies}
}
\description{
\code{PlotStationEra} plots the station values together with the
  corresponding ERA20C, ERA-I, and HErZ pixel and provides the correlation
  between these time series. Optionally, it is possible to plot the anomaly.
  The plot is saved in pdf format and there is no return value.
}

