% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PlotStationEraSeasons}
\alias{PlotStationEraSeasons}
\title{Plot seasonal time series of station data against locally corresponding
  global and reginal reanalyses}
\usage{
PlotStationEraSeasons(Era20cXts, EraIXts, HerzXts, StatXts, titname, outdir,
  fname, width, height, anomaly = FALSE, seasons = FALSE)
}
\arguments{
\item{Era20cXts}{monthly mean extended time series of the ERA20C pixel
corresponding to the station location}

\item{EraIXts}{same as above for ERA-Interim}

\item{HerzXts}{same as above for HErZ}

\item{StatXts}{monthly mean extended time series of the station values}

\item{titname}{string of the plot title name}

\item{outdir}{string of the output directory into which the plot is saved}

\item{fname}{string of the file name of the plot}

\item{width,height}{of the plot in inches}

\item{anomaly}{is an optional parameter which determines whether to plot anomalies}

\item{seasons}{is an optional parameter which determines whetther to plot the
monthly (F) or seasonal (T) values of the above time series}
}
\description{
THIS FUNCTION IS NOT YET FINISHED.
  \code{PlotStationEraSeasons} plots seasonal means of station values against
  locally corresponding time series (pixels) of global and regional reanalyses. The
  rolling mean and optionally the anomalies are plotted. The plot is saved in pdf
  format and there is no return value.
}

