% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PlotStationEraSelMonths}
\alias{PlotStationEraSelMonths}
\title{Prepare and plot specifically selected months of station data and ERA20C,
  ERA-I, and COSMO-REA6 data.}
\usage{
PlotStationEraSelMonths(Era20cXts, EraIXts, HerzXts, StatXts, titname, outdir,
  fname, anomaly = FALSE)
}
\arguments{
\item{Era20cXts}{monthly mean extended time series of the ERA20C pixel
corresponding to the station location}

\item{EraIXts}{same as above for ERA-Interim}

\item{HerzXts}{same as above for COSMO-REA6}

\item{StatXts}{monthly mean extended time series of the station values}

\item{titname}{string of the plot title name}

\item{outdir}{string of the output directory into which the plot is saved}

\item{fname}{is a string holding file name of the pdf plot to be created.}

\item{anomaly}{is an optional parameter which determines whether to plot anomalies}
}
\description{
\code{PlotStationEraSelMonths} prepares extended time series of
  station data, ERA20C, ERA-I, and COSMO-REA6 data - for specific months only. These
  months are set hard-coded within this function. Of course, this setting can be
  put into the Settings.R file if necessary.
  Optionally, it is possible to prepare the anomalies. A generic plotting routine
  is called which actually performs the plotting.
}
\note{
need to adopt titname to months; need to plot into four different panals
}

