% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PlotStationEraSelSeasons}
\alias{PlotStationEraSelSeasons}
\title{Prepare and plot seasonal time series of station data and ERA20C, ERA-I,
  and COSMO-REA6 data.}
\usage{
PlotStationEraSelSeasons(Era20cXts, EraIXts, HerzXts, StatXts, titname, outdir,
  fname, anomaly = FALSE, seasons = FALSE)
}
\arguments{
\item{Era20cXts}{monthly mean extended time series of the ERA20C pixel
corresponding to the station location}

\item{EraIXts}{same as above for ERA-Interim}

\item{HerzXts}{same as above for COSMO-REA6}

\item{StatXts}{monthly mean extended time series of the station values}

\item{titname}{string of the plot title name}

\item{outdir}{string of the output directory into which the plot is saved}

\item{fname}{is a string holding file name of the pdf plot to be created.}

\item{anomaly}{is an optional parameter which determines whether to plot anomalies}

\item{seasons}{is an optional parameter which determines whetther to plot the
monthly (F) or seasonal (T) values of the above time series}
}
\description{
\code{PlotStationEraSelSeasons} prepares extended time series of
  seasonal means, and optionally their anomalies, of station data and locally
  corresponding global and regional reanalyses. The seasons for which data shall
  be prepared and plotted need to be set hard-coded within this function. Of
  course, this setting can be put into the Settings.R file if necessary.
  A generic plotting routine is called which actually performs the plotting.
}

