% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PlotStationEraSQ}
\alias{PlotStationEraSQ}
\title{Prepare and plot time series of station data and ERA20C, ERA-I,
  and HErZ data.}
\usage{
PlotStationEraSQ(Era20cXts, EraIXts, HerzXts, StatXts, titname, outdir, fname,
  era.monthly = FALSE)
}
\arguments{
\item{Era20cXts}{extended time series of the ERA20C pixel corresponding to the
station location}

\item{EraIXts}{same as above for ERA-Interim}

\item{HerzXts}{same as above for HErZ}

\item{StatXts}{extended time series of the station values}

\item{titname}{string of the plot title name}

\item{outdir}{string of the output directory into which the plot/s is/are saved}

\item{fname}{string of the output file name}

\item{era.monthly}{is an optional boolean which determines whether data passed
is monthly (T) or daily (F) data. The default value is to use daily data
(era.monthly=FALSE).}
}
\description{
This function plots station and reanalysis data into scatter plots
  and Quantile-quantile plots (QQ-plot).
}

