% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PlotTowerERAprofileAnnualCycle}
\alias{PlotTowerERAprofileAnnualCycle}
\title{Plot annual cycle between tower measurements and reanalysis data.}
\usage{
PlotTowerERAprofileAnnualCycle(tower.obj, fname)
}
\arguments{
\item{fname}{string of the file name of the plot}

\item{tower.df}{is a data frame holding the data of tower measurements and
corresponding reanalysis data}
}
\description{
Plot the annual cycle between tower measurements and reanalysis data
  at all available heights for absolute and relative values.  Tower measurements
  of Lindenberg, Cabauw, Fino1, and Fino2 are supported.
}

