% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PlotTowerERAprofileAnnualVar}
\alias{PlotTowerERAprofileAnnualVar}
\title{Plot annual variability between tower measurements and reanalysis data.}
\usage{
PlotTowerERAprofileAnnualVar(tower.obj, fname)
}
\arguments{
\item{tower.obj}{is a ClimObject holding the data of tower measurements and
corresponding reanalysis data}

\item{fname}{is a string holding file name of the pdf plot to be created.}
}
\description{
Plot the annual variability between tower measurements and reanalysis
  data at the heights of 10m and 100m where applicable. Two months are selected
  to visualize the annual variability; January and August are predifined and may
  be changed but only locally so far. Tower measurements of Lindenberg, Cabauw,
  Fino1, and Fino2 are supported.
}

