% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PlotTowerERAprofileBP}
\alias{PlotTowerERAprofileBP}
\title{Plot wind speed profile as box plots of tower measurements and ERA.}
\usage{
PlotTowerERAprofileBP(tower.obj, fname, ana.time.res)
}
\arguments{
\item{tower.obj}{is a ClimObject holding the data of tower measurements and
corresponding reanalysis data}

\item{fname}{is a string holding file name of the pdf plot to be created.}

\item{ana.time.res}{named list holding parameters monthly="monthly",
daily="daily", hourly="hourly", and time.res= to determine the time resolution
of the data to be read.}
}
\description{
This function plots the wind speed profile of tower measurements and
  data from reanalyses as box plots. Tower measurements of Lindenberg, Cabauw,
  Fino1, and Fino2 are supported.
}

