% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PlotTowerERAprofileRelDiff}
\alias{PlotTowerERAprofileRelDiff}
\title{Plot relative difference between tower measurements and reanalysis data.}
\usage{
PlotTowerERAprofileRelDiff(tower.df, tower.name, fname)
}
\arguments{
\item{tower.df}{is a data frame holding the data of tower measurements and
corresponding reanalysis data}

\item{tower.name}{is a string holding the tower name}

\item{fname}{string of the file name of the plot}
}
\description{
Plot the relative difference between tower measurements at 10m and
  100m compared to reanalysis data. Tower measurements of Lindenberg, Cabauw,
  Fino1, and Fino2 are supported.
}

