% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ReadEraNetcdf2Xts}
\alias{ReadEraNetcdf2Xts}
\title{Read monthly or daily ERA20C or ERA-Interim data and convert to xts}
\usage{
ReadEraNetcdf2Xts(era.param, era.fname, era.tsstart, era.tsend, lonidx, latidx,
  ana.time.res, era20c = TRUE, verb.dat = FALSE)
}
\arguments{
\item{era.param}{string of length n holding the parameters of the n different
model levels (one for ERA-I, two for ERA20C).}

\item{era.fname}{string holding the file name of the ERA data.}

\item{era.tsstart}{string of the start date of the ERA data of the format
c(YYYY,MM).}

\item{era.tsend}{same as above for the end date.}

\item{lonidx}{numeric value of the longitude station location which to extract
off the ERA grid.}

\item{latidx}{same as above for latitude.}

\item{ana.time.res}{named list holding parameters monthly="monthly",
daily="daily", hourly="hourly", and time.res= to determine the time resolution
of the data to be read.}

\item{era20c}{boolean to determine whether the data is ERA20C (T) or
ERA-Interim (F).}

\item{verb.dat}{optional boolean to determine whether to print out what's going
on (T). The default is to suppress printing (verb.dat = FALSE).}
}
\value{
A named list holding the extended time series of the ERA data at 10m
  height for ERA-Interim and 10m and 100m for ERA20C.: (era10=,era20c100=). If
  ERA-Interim data is read, a NULL value will be returned for era20c100.
}
\description{
This function reads monthly or daily Era data at the location of the
  station location. The time steps of the monthly time series is given with
  precision of months, those of daily time series with precision of days. The
  return values are Era 10m windspeed and in case of ERA20C data windspeed in
  100m height.
}

