% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readGrib}
\alias{readGrib}
\title{Reading grib files based on wgrib.}
\usage{
readGrib(filename, nlon, nlat, nlev, var = "undef", out = "Rfile.dat",
  recs = c(0), pipe = "", verb.grib = FALSE)
}
\arguments{
\item{filename}{string of the grib file name to read from}

\item{nlon,nlat,nlev}{specify the extent of the data in number of longitude,
latitude and vertical level steps (pixels)}

\item{var}{string of the variable name to read}

\item{out}{a temporary file into which output is dumped. NOTE: this needs to be
specifyied if this function is called more than once in parralel (i.e., totally
independent from each other)}

\item{recs}{up to two concatinated integers which specify the beginning and end
of the records (time steps) to read}

\item{pipe}{string to optionally specify additional selection chriterium. Default
is an empty string.}

\item{verb.grib}{optional boolean to determine whether to verbosely write out
what's happening (T) or not (F); the default is to suppress output
(verb.grib=FALSE).}
}
\value{
x the array containing the read data
}
\description{
\code{readGrib} is based on wgrib to read grib files. It can read
  files by parameter and by specifyed level and time step. This function has been
  written by Chris Bollmeyer of Uni Bonn, received in 2014-09-25
}

