% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ReadNetcdf}
\alias{ReadNetcdf}
\title{Reading netCDF files}
\usage{
ReadNetcdf(variable, infile, start = NULL, count = NULL, revert = FALSE,
  verb.dat = FALSE)
}
\arguments{
\item{variable}{is a string containing the variable name to be read of the nc-file}

\item{infile}{is a string holding the file name to read}

\item{start}{optional parameter to set the beginning of the variable array to be
read. It needs to be of the same size as variable (for further details see the
help of \code{\link{get.var.ncdf}}). The default is to read the complete
variable which leaves start unset.}

\item{count}{optional parameter to set the size of the variable array to be read.
It needs to be of the same size as variable (for further details see the help
of \code{\link{get.var.ncdf}}). The default is to read the complete variable
which leaves count unset.}

\item{revert}{in an optional boolean to decided whether to revert the data in
North-South direction. Default is not to revert the data (FALSE)}

\item{verb.dat}{optional boolean to determine whether to verbosely write out what
is happening (T) or not (F); the default is to suppress output (verb.dat=FALSE).}
}
\value{
Return a named list (data=,lon=,lat=,time=) holding the read data,
  longitude, latitude, and time values.
}
\description{
\code{ReadNetcdf} reads nc-files provided a variable name and the
  file name to read from. Longitude, latitude, and time values are read if
  available. An optional paramter "revert" regulates whether the data values
  should be reverted in North-South. The variable can be read partially depending
  on the provided start and count parameters, or completely if they are omitted.
}

