% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ReadNetcdfLonLat}
\alias{ReadNetcdfLonLat}
\title{Read longitude and latitude arrays off a netCDF file.}
\usage{
ReadNetcdfLonLat(infile)
}
\arguments{
\item{infile}{string of the file name to be read}
}
\value{
Return a named list (lon=,lat=) for the longitude and latitude arrays
  (1D or 2D) read off the netCDF file.
}
\description{
\code{ReadnercdfLonLat} is a shortcut to only read the longitude and latitude
  information off a netCDF file. If they don't exist a NULL value will be returned.
}

