% Generated by roxygen2 (4.0.2): do not edit by hand
\name{RelDiff}
\alias{RelDiff}
\title{Calculate the relative difference between two values.}
\usage{
RelDiff(value, mean.value)
}
\arguments{
\item{value}{is the value of which the relvative difference is to be calcualted
compated to mean.value}

\item{mean.value}{is the mean value from above}
}
\value{
is the relative difference
}
\description{
Provided is a value and the mean value to which the relative
  difference is to be computed.
}

