% Generated by roxygen2 (4.0.2): do not edit by hand
\name{RemoveFiles}
\alias{RemoveFiles}
\title{A function to remove files.}
\usage{
RemoveFiles(extension = NULL)
}
\arguments{
\item{extension}{is a (vector of) string(s) which holds the extensions to of the
files to be removed.}
}
\description{
\code{RemoveFiles} can remove any files with provided extensions;
  by defaul it explicitely removes *.txt and *.html files.
}

