% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scatterPlot}
\alias{scatterPlot}
\title{Produce a scatter plot}
\usage{
scatterPlot(X, Y, yliml, ylimh, titname, xlabname, ylabname, text.str = NULL,
  xaxis = "s", yaxis = "s", plot.col = "blue")
}
\arguments{
\item{X}{first data sample for the scatter plot of class numeric}

\item{Y}{second data sample}

\item{yliml}{lower numeric bond of the y-axis limits}

\item{ylimh}{high numeric bond of the y-axis limits}

\item{titname}{string of the title name of the plot}

\item{xlabname}{string for the name of the x-axis label}

\item{ylabname}{string for the name of the y-axis label}

\item{text.str}{named variable of type string which will be written into the
plot with \code{text}.}

\item{xaxis,yaxis}{are two named variables which set whether to plot the x- and
y-axis (xaxis='s', yaxis='s') or omit either one (either set to 'n'). It not
set the default value will be used which is set to 's' meaning that the axis
will be plotted.}

\item{plot.col}{is a string holding the plotting color which is set by default to
"blue".}
}
\description{
Prodcue a scatter plot of two data set samples.
}

