% Generated by roxygen2 (4.0.2): do not edit by hand
\name{SetToDate}
\alias{SetToDate}
\title{Determine start/end string to set extended time series.}
\usage{
SetToDate(tsstart, tsend)
}
\arguments{
\item{tsstart}{numeric of the format c(YYYY,M) holding the start date}

\item{tsend}{as above holding the end date}
}
\value{
timestr string of the format YYYYM/YYYYM holding start date/end date.
}
\description{
A start and end date in the format of c(YYYY,M) are put together in
  a string of the format YYYYM/YYYYM.
}

