% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ExtractERAxts}
\alias{ExtractERAxts}
\title{Extract ERA pixel corresponding to station location.}
\usage{
ExtractERAxts(data, time.vals, lon, lat, tsstart, tsend, stat.lon, stat.lat)
}
\arguments{
\item{data}{data set (reanalysis) from which to extract the time series at the
station location}

\item{time.vals}{time period covered by the data set}

\item{lon}{longitude grid points of the data set}

\item{lat}{latitude grid points of the data set}

\item{tsstart}{character string of the format (YYYY,M) of the start date of the
data set time series}

\item{tsend}{character string as above of the end date of the data set}

\item{stat.lon}{station longitude}

\item{stat.lat}{station latitude}
}
\value{
era.xts[timestr] is the time series extracted off the data set at the
  pixel corresponding to the provided station location
  Extract nearest pixel of reanalysis corresponding to station lon, lat.
}
\description{
\code{ExtractERAxts} extracts the nearest pixel of a provided data
  set (reanalysis) given the station location (lon, lat).
}

