% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Plot100mEraHerz}
\alias{Plot100mEraHerz}
\title{Compare 100m wind speed of ERA20C and COSMO-REA6 pixel by pixel.}
\usage{
Plot100mEraHerz(Era20cXts, HerzXts, titname, statname, outdir, fname,
  ana.time.res)
}
\arguments{
\item{Era20cXts}{extended time series of an ERA20C pixel}

\item{HerzXts}{same as above for COSMO-REA6}

\item{titname}{string of the plot title name}

\item{statname}{string of the station name whose pixel is plotted}

\item{outdir}{string of the output directory into which the plot is saved}

\item{fname}{is a string holding file name of the pdf plot to be created.}

\item{ana.time.res}{named list holding parameters monthly="monthly",
daily="daily", hourly="hourly", and time.res= to determine the time resolution
of the data to be read.}
}
\description{
\code{Plot100mEraHerz} compares the 100m wind speed of the ERA20C
  global reanalysis with the 116m wind speed of the COSMO-REA6 regional reanalysis.
  This function performs a pixel wise comparison at the station locationto
  provided by the package. Scatter plots, QQplots, histogram plots, and the
  PDFscore are produced.
}

