% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Plot100mEraHerz}
\alias{Plot100mEraHerz}
\title{Compare 100m wind speed of ERA20C and HErZ pixel by pixel.}
\usage{
Plot100mEraHerz(Era20cXts, HerzXts, titname, statname, outdir, fname,
  era.monthly, width, height)
}
\arguments{
\item{Era20cXts}{extended time series of an ERA20C pixel}

\item{HerzXts}{same as above for HErZ}

\item{titname}{string of the plot title name}

\item{statname}{string of the station name whose pixel is plotted}

\item{outdir}{string of the output directory into which the plot is saved}

\item{fname}{string of the file name of the plot}

\item{era.monthly}{is an optional boolean which determines whether data passed
is monthly (T) or daily (F) data. The default value is to use daily data
(era.monthly=FALSE).}

\item{width,height}{of the plot in inches}
}
\description{
\code{Plot100mEraHerz} compares the 100m wind speed of the ERA20C
  global reanalysis with the 116m wind speed of the HErZ regional reanalysis.
  This function performs a pixel wise comparison at the station locationto
  provided by the package. Scatter plots, QQplots, histogram plots, and the
  PDFscore are produced.
}

