% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PlotHistograms}
\alias{PlotHistograms}
\title{Plot the wind speed histograms of station data and reanalyses at different
  heights.}
\usage{
PlotHistograms(outdir, fname, station.name, ana.time.res, Era20cXts10 = NULL,
  Era20cXts100 = NULL, EraIXts = NULL, HerzXts10 = NULL,
  HerzXts35 = NULL, HerzXts69 = NULL, HerzXts116 = NULL,
  HerzXts178 = NULL, HerzXts258 = NULL, StatXts = NULL,
  plot.10m = FALSE, plot.100m = FALSE, plot.HerzProfile = FALSE)
}
\arguments{
\item{outdir}{is a string containing the output path of the plot}

\item{fname}{is a string holding file name of the pdf plot to be created.}

\item{station.name}{is a string holding the station name at which location the
distribution data was extraced}

\item{ana.time.res}{named list holding parameters monthly="monthly",
daily="daily", hourly="hourly", and time.res= to determine the time resolution
of the data to be read.}

\item{Era20cXts10,Era20cXts100,EraIXts}{extended time series of ERA20C and ERA-I
10m and 100m height}

\item{HerzXts10,HerzXts35,HerzXts69,HerzXts116,HerzXts178,HerzXts258}{extended
time series of available HErZ data at different heights}

\item{StatXts}{extended time series of station data}

\item{plot.10m,plot.100m,plot.HerzProfile}{optional parameters setting whether
to plot only 10m values, only 100m values, or the herz profile, respectively.
The default value is FALSE for all three settings.}
}
\description{
The distributions of monthly and daily wind speed station data and
  reanalysis data, optionally at differnet heights are plotted into histograms.
}

