% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PlotHistogramsTower}
\alias{PlotHistogramsTower}
\title{Plot histrograms for tower measurements.}
\usage{
PlotHistogramsTower(outdir, fname, ana.time.res, tower.obj, plot.10m = FALSE,
  plot.100m = FALSE, plot.HerzProfile = FALSE)
}
\arguments{
\item{outdir}{is a string containing the output path of the plot}

\item{fname}{is a string holding file name of the pdf plot to be created.}

\item{ana.time.res}{named list holding parameters monthly="monthly",
daily="daily", hourly="hourly", and time.res= to determine the time resolution
of the data to be read.}

\item{tower.obj}{is a ClimObject holding the data of tower measurements and
corresponding reanalysis data}
}
\description{
Plot histograms for the Lindenberg, Cabauw, Fino1 and Fino2 tower
  tower measurements. Histograms are plotted for 10m and 100m heights; seperately
  and in comparison with reanalysis data.
}

