% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PlotStationEra}
\alias{PlotStationEra}
\title{Plot station measurements together with ERA20C, ERA-I, and HErZ.}
\usage{
PlotStationEra(Era20cXts, EraIXts, HerzXts, StatXts, titname, outdir, fname,
  monthly = TRUE, anomaly = FALSE)
}
\arguments{
\item{Era20cXts}{monthly mean extended time series of the ERA20C pixel
corresponding to the station location}

\item{EraIXts}{same as above for ERA-Interim}

\item{HerzXts}{same as above for HErZ}

\item{StatXts}{monthly mean extended time series of the station values}

\item{titname}{string of the plot title name}

\item{outdir}{string of the output directory into which the plot is saved}

\item{fname}{string of the file name of the plot}

\item{monthly}{is an optional parameter which determines to plot the monthly
values of the above time series}

\item{anomaly}{is an optional parameter which determines whether to plot anomalies}
}
\description{
\code{PlotStationEra} plots the station values together with the
  corresponding ERA20C, ERA-I, and HErZ pixel and provides the correlation
  between these time series. Optionally, it is possible to plot the anomaly.
  The plot is saved in pdf format and there is no return value.
}

