% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PlotStationEraMonths}
\alias{PlotStationEraMonths}
\title{Plot station measurements together with ERA20C, ERA-I, and HErZ.}
\usage{
PlotStationEraMonths(Era20cXts, EraIXts, HerzXts, StatXts, titname, outdir,
  fname, width, height, anomaly = FALSE)
}
\arguments{
\item{Era20cXts}{monthly mean extended time series of the ERA20C pixel
corresponding to the station location}

\item{EraIXts}{same as above for ERA-Interim}

\item{HerzXts}{same as above for HErZ}

\item{StatXts}{monthly mean extended time series of the station values}

\item{titname}{string of the plot title name}

\item{outdir}{string of the output directory into which the plot is saved}

\item{width,}{height of the plot in inches}

\item{anomaly}{is an optional parameter which determines whether to plot anomalies}

\item{monthly}{is an optional parameter which determines to plot the monthly
values of the above time series}
}
\description{
THERE NEEDS TO BE A DIFFERENCE TO \code{\link{PlotStationEra}}.
  CHECK WHICH IT IS !!!
  \code{PlotStationEra} plot the station values together with the
  corresponding ERA20C, ERA-I, and HErZ pixel and provides the correlation between
  these time series. Optionally, it is possible to plot the anomaly. The plot is
  saved in pdf format and there is no return value. The plot is
  saved in pdf format and there is no return value.
}

