% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PlotStationEraSQ}
\alias{PlotStationEraSQ}
\title{Prepare and plot time series of station data and ERA20C, ERA-I,
  and HErZ data.}
\usage{
PlotStationEraSQ(Era20cXts, EraIXts, HerzXts, StatXts, titname, outdir, fname,
  ana.time.res)
}
\arguments{
\item{Era20cXts}{extended time series of the ERA20C pixel corresponding to the
station location}

\item{EraIXts}{same as above for ERA-Interim}

\item{HerzXts}{same as above for HErZ}

\item{StatXts}{extended time series of the station values}

\item{titname}{string of the plot title name}

\item{outdir}{string of the output directory into which the plot/s is/are saved}

\item{fname}{is a string holding file name of the pdf plot to be created.}

\item{ana.time.res}{named list holding parameters monthly="monthly",
daily="daily", hourly="hourly", and time.res= to determine the time resolution
of the data to be read.}
}
\description{
This function plots station and reanalysis data into scatter plots
  and Quantile-quantile plots (QQ-plot).
}

