% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PlotTowerERAprofileAnnualVar}
\alias{PlotTowerERAprofileAnnualVar}
\title{Plot annual variability between tower measurements and reanalysis data.}
\usage{
PlotTowerERAprofileAnnualVar(tower.df, tower.name, fname, width, height)
}
\arguments{
\item{tower.df}{is a data frame holding the data of tower measurements and
corresponding reanalysis data}

\item{tower.name}{is a string holding the tower name}

\item{fname}{string of the file name of the plot}

\item{width,height}{of the plot in inches}
}
\description{
Plot the annual variability between tower measurements and reanalysis
  data at the heights of 10m and 100m where applicable. Two months are selected
  to visualize the annual variability; January and August are predifined and may
  be changed but only locally so far. Tower measurements of Lindenberg, Cabauw,
  Fino1, and Fino2 are supported.
}

