% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PlotTowerERAprofileBP}
\alias{PlotTowerERAprofileBP}
\title{Plot wind speed profile as box plots of tower measurements and ERA.}
\usage{
PlotTowerERAprofileBP(tower.df, tower.name, fname, era.monthly)
}
\arguments{
\item{tower.df}{is a data frame holding the data of tower measurements and
corresponding reanalysis data}

\item{tower.name}{is a string holding the tower name}

\item{fname}{string of the file name of the plot}

\item{era.monthly}{is an optional boolean which determines whether data passed
is monthly (T) or daily (F) data. The default value is to use daily data
(era.monthly=FALSE).}
}
\description{
This function plots the wind speed profile of tower measurements and
  data from reanalyses as box plots. Tower measurements of Lindenberg, Cabauw,
  Fino1, and Fino2 are supported.
}

