% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ReadNetcdf}
\alias{ReadNetcdf}
\title{Reading netCDF files.}
\usage{
ReadNetcdf(variable, infile, revert = FALSE)
}
\arguments{
\item{variable}{is a string containing the variable name to be read of the nc-file}

\item{infile}{is a string holding the file name to read}

\item{revert}{in an optional boolean to decided whether to revert the data in
North-South direction. Default is not to revert the data (FALSE)}
}
\value{
data,lon,lat,time.vals in a list holding the read data, longitude,
  latitude, and time values
}
\description{
\code{ReadNetcdf} read nc-files provided a variable name and the
  file name to read from. Longitude, latitude, and time values are read if
  available. An optional paramter "revert" regulates whether the data values
  should be reverted in North-South.
}

